Move form consent to new component
